Fixed license race condition.

+++ batch_processing/get_license.py
@@ -5,22 +5,46 @@ import socket
 import click
 import portalocker
 
-def get_license_string():
+def register_license(hostname, server, license_string):
+    command = f'grbgetkey {license_string} --verbose --path licenses/{hostname} '
+    if server is not None:
+        command += f'--server {server}'
+
+    os.system(command)
+
+def retrieve_license(hostname, server, license_path):
     print('Waiting for lock...')
+
     with portalocker.Lock('license_list.txt', 'r+') as fh:
-        licenses = [l.strip() for l in fh.readlines()]
+        print('Lock acquired.')
 
-        if len(licenses) == 0:
-            return None
+        # Check that the license hasn't already been registered during waiting
+        if license_path.exists():
+            # If the license has already been registered, return
+            return
+        else:
+            # If the license has not been registered, register it
+            licenses = [l.strip() for l in fh.readlines()]
 
-        print('Acquiring license', licenses[0])
-        fh.seek(0)
-        fh.write('\n'.join(licenses[1:]))
-        fh.truncate()
+            if len(licenses) == 0:
+                # No licenses, abort
+                exit(1)
 
-    print('Released lock')
+            license = licenses[0]
 
-    return licenses[0]
+            print('Acquired license', license)
+
+            print('Registering...')
+            register_license(hostname, server, license)
+            print('Registered license', license)
+
+            fh.seek(0)
+            fh.write('\n'.join(licenses[1:]))
+            fh.truncate()
+
+        print('Released lock')
+
+        return licenses[0]
 
 @click.command()
 @click.option('--server', type=str, default=None)
@@ -31,17 +55,9 @@ def main(server):
     hostname =  socket.gethostname()
 
     license_path = base_folder / hostname / 'gurobi.lic'
+
     if not license_path.exists():
-        license_string = get_license_string()
-
-        if license_string is None:
-            exit(1)
-
-        command = f'grbgetkey {license_string} --verbose --path licenses/{hostname} '
-        if server is not None:
-            command += f'--server {server}'
-
-        os.system(command)
+        retrieve_license(hostname, server, license_path)
 
 if __name__ == '__main__':
     main()
